make changeDurability mechanic method easier to access
